fix: use bool directly in patch_epic (rusqlite implements ToSql for bool)

+++ src/db/queries.rs
@@ -546,7 +546,7 @@ impl super::EpicCrud for Database {
         }
         if let Some(ad) = patch.auto_dispatch {
             sets.push("auto_dispatch = ?");
-            values.push(Box::new(ad as i64));
+            values.push(Box::new(ad));
         }
 
         sets.push("updated_at = datetime('now')");
